+++ js/wallets.js
@@ -1,0 +1,6 @@
+function init() {
+    console.log('wallets.js loaded')
+}
+
+exports.init = init
+
